fix(kapi): kure neden dustu — SAVUNMA VARDI, VERIYI HIC GORMEDI
Kure capasi LM ile kostu ve Cd 0,142 verdi (referans 0,47, sapma ~%70).
"Model yanlis" degil: modelin istedigi ag hic kurulmamis. Zincir kendi
verisinden okunuyor (validation_anchors_runs/_anchor_sphere/sonuc.json):

  10 katman istendi -> ORTALAMA 0,535 oruldu
  -> y+ ort 59 (tepe 214), hedef <=1
  -> LM duvar-fonksiyonu aginda kostu -> laminer altkatman hic gorulmedi

Iki kusur bunu gorunmez kilmisti:

1. KATMAN AYRISTIRICISI ONDALIK OKUMUYORDU. snappyHexMesh yuz basina
   ORTALAMA katman yazar ("_anchor_sphere_prep 1728 0.535 9.21e-05 13.9"),
   tam sayi degil. `int(t[2])` ValueError atiyor, TUM olcum dusuyor ve
   sonuc "olculemedi" oluyordu. Depoda katman-cokmesi hukmu (katman_hukmu
   -> "COKTU") ZATEN VARDI; ayristirici once dustugu icin veriyi hic
   gormuyordu. Ayni imza Ahmed'de de olculdu.

2. DUVAR KAPISI MODELDEN HABERSIZDI. y+ 30-300 kOmegaSST icin mesru,
   gecis modeli icin gecersiz. duvar_hukmu ayrim yapmadigi icin ayni agi
   "duvar fonksiyonu bandinda" diye GECIRIYORDU. gecis_modeli_onkosulu
   ISTEGI denetler (n_layers=10, y+=1 -> tamam); GERCEKLESENI denetleyen
   yoktu. duvar_hukmu artik model aliyor ve capa kosucusu geciriyor.

Olculdu: ayni agla kOmegaSST -> True, kOmegaSSTLM -> False
  ("kOmegaSSTLM DUVAR-COZUNUR ag ister (y+<=5) ama olculen y+ ortalamasi 59,1")

Kure artik yanlis sayiyi vermek yerine GEREKCELI reddediyor. Katmanlarin
gercekten orulmesi ayri bir is; artik bilinmeyen degil, olculmus hedef.

Paket: 1645 gecti, 2 atlandi (diskte olmayan fixture).

+++ vehicle_pipeline.py
@@ -695,7 +695,18 @@ def parse_layer_report(log: Path) -> dict:
         if len(t) < 5:
             break
         try:
-            out["yamalar"].append({"ad": t[0], "yuz": int(t[1]), "katman": int(t[2]),
+            # KATMAN SAYISI ONDALIKTIR. snappyHexMesh yüz başına ORTALAMA
+            # katmanı yazar (ör. "0.535"), tam sayı değil. `int(t[2])`
+            # ValueError atıyor ve TÜM ölçüm düşüyordu — yani katmanların
+            # örülmediği durumda araç bunu SÖYLEYEMİYORDU.
+            #
+            # Ölçüldü (2026-08-19, küre çapası): 10 katman istendi, ortalama
+            # 0,535 örüldü, y⁺ 59 çıktı (hedef ≤1). Depoda katman çökmesini
+            # yakalayan bir hüküm ZATEN var (`katman_hukmu` → "COKTU") ama
+            # ayrıştırıcı önce düştüğü için o savunma veriyi HİÇ görmüyordu.
+            # Aynı imza Ahmed çapasında da vardı.
+            out["yamalar"].append({"ad": t[0], "yuz": int(t[1]),
+                                   "katman": float(t[2]),
                                    "kalinlik_m": float(t[3]),
                                    "kalinlik_pct": float(t[4])})
         except ValueError:
